+++ packages/client/src/common/commonTypes.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import type { BaseLanguageClient } from 'vscode-languageclient/browser.js';
+import type { BaseLanguageClient } from 'vscode-languageclient/browser';
 
 export type ConnectionConfigOptions =
   | WebSocketConfigOptionsDirect