+++ packages/vscode-ws-jsonrpc/src/index.ts
@@ -3,7 +3,10 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-export * from './disposable.js';
-export * from './socket/index.js';
-export * from './logger.js';
-export * from './connection.js';
+export * from './common/connection.js';
+export * from './common/disposable.js';
+export * from './common/logger.js';
+export * from './common/types.js';
+export * from './socket/connection.js';
+export * from './socket/reader.js';
+export * from './socket/writer.js';