+++ verify/next/app/langium-dsl/worker/langium-server.ts
@@ -7,7 +7,7 @@
 import { EmptyFileSystem } from 'langium';
 import { startLanguageServer } from 'langium/lsp';
 import { createLangiumGrammarServices } from 'langium/grammar';
-import { BrowserMessageReader, BrowserMessageWriter, createConnection } from 'vscode-languageserver/browser.js';
+import { BrowserMessageReader, BrowserMessageWriter, createConnection } from 'vscode-languageserver/browser';
 
 /* browser specific setup code */
 export let messageReader: BrowserMessageReader | undefined;