+++ verify/next/vite.config.ts
@@ -30,7 +30,7 @@ export const config = defineConfig({
       'vscode-jsonrpc',
       'vscode-languageclient',
       'vscode-languageserver',
-      'vscode-languageserver/browser.js',
+      'vscode-languageserver/browser',
       'vscode-languageserver-protocol'
     ]
   },